rename Game -> Global in examples
