+++ examples/03-tiles/Global.js
