+++ examples/03-tiles/index.html
@@ -72,7 +72,7 @@
         
 
 <script type="module">
-import Game             from './Game.js'
+import Global             from './Global.js'
 import * as Cobalt      from '../../bundle.js'
 import constants        from './constants.js'
 import dat              from 'https://cdn.skypack.dev/pin/dat.gui@v0.7.9-2wtQAdFH5SRwnJLDWGNz/mode=imports,min/optimized/dat.gui.js'
@@ -89,10 +89,10 @@ async function main () {
 
     const viewportWidth = constants.GAME_WIDTH
     const viewportHeight = constants.GAME_HEIGHT
-    Game.renderer = await Cobalt.init(canvas, viewportWidth, viewportHeight)
+    Global.renderer = await Cobalt.init(canvas, viewportWidth, viewportHeight)
 
     // instantiate all resource nodes
-    const tileAtlasNode = await Cobalt.initNode(Game.renderer, {
+    const tileAtlasNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:tileAtlas',
         refs: { },
         options: {
@@ -103,7 +103,7 @@ async function main () {
         }
     })
 
-    const hdrTex = await Cobalt.initNode(Game.renderer, {
+    const hdrTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -115,7 +115,7 @@ async function main () {
         }
     })
 
-    const emissiveTex = await Cobalt.initNode(Game.renderer, {
+    const emissiveTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -127,7 +127,7 @@ async function main () {
         }
     })
 
-    const bloomTex = await Cobalt.initNode(Game.renderer, {
+    const bloomTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -139,7 +139,7 @@ async function main () {
         }
     })
 
-    const bloomNode = await Cobalt.initNode(Game.renderer, {
+    const bloomNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:bloom',
         refs: {
             // key is the var name defined in this node
@@ -167,7 +167,7 @@ async function main () {
         }
     })
 
-    const tmpTex = await Cobalt.initNode(Game.renderer, {
+    const tmpTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -179,7 +179,7 @@ async function main () {
         }
     })
 
-     const tileNodeBG = await Cobalt.initNode(Game.renderer, {
+     const tileNodeBG = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
@@ -195,7 +195,7 @@ async function main () {
     })
 
     // instantiate all runnable nodes
-    const tileNodeBg = await Cobalt.initNode(Game.renderer, {
+    const tileNodeBg = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
@@ -209,7 +209,7 @@ async function main () {
         }
     })
 
-    const tileNodeFg = await Cobalt.initNode(Game.renderer, {
+    const tileNodeFg = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
@@ -224,7 +224,7 @@ async function main () {
     })
 
 
-    const compositeNode = await Cobalt.initNode(Game.renderer, {
+    const compositeNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:composite',
         refs: {
             hdr: hdrTex,
@@ -235,7 +235,7 @@ async function main () {
     })
 
 
-    const b = await Cobalt.initNode(Game.renderer, {
+    const b = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbBlit',
         refs: {
             in: tmpTex,
@@ -246,11 +246,11 @@ async function main () {
 
 
     // use resizeViewport to init values on load:
-    resizeViewport(Game.renderer, window.innerWidth, window.innerHeight)
+    resizeViewport(Global.renderer, window.innerWidth, window.innerHeight)
 
     // window resize is *expensive* - best to debounce:
     const debouncedResize = debounce(function () {
-       resizeViewport(Game.renderer, window.innerWidth, window.innerHeight) 
+       resizeViewport(Global.renderer, window.innerWidth, window.innerHeight) 
     }, 50)
 
     window.addEventListener('resize', debouncedResize, { passive: true })
@@ -290,9 +290,9 @@ async function main () {
 
     
     const world = ECS.createWorld()
-    ECS.addSystem(world, rendererSystem(Game.renderer))
+    ECS.addSystem(world, rendererSystem(Global.renderer))
 
-    Cobalt.setViewportPosition(Game.renderer, [ 240, 135 ])
+    Cobalt.setViewportPosition(Global.renderer, [ 240, 135 ])
     
     ECS.addSystem(world, function cameraLoopSystem(world) {
         const onUpdate = function (dt) {
@@ -300,19 +300,19 @@ async function main () {
             const x = (Math.sin(elapsed / 2000) * 0.5 + 0.5) * 128
             const y = (Math.sin(elapsed / 5000) * 0.5 + 0.5) * 170
 
-            //const { tileScale } = Game.renderer.tile
+            //const { tileScale } = Global.renderer.tile
 
-            Cobalt.setViewportPosition(Game.renderer, [ x, y ])
+            Cobalt.setViewportPosition(Global.renderer, [ x, y ])
         }
 
         return { onUpdate }
     })
 
-    Game.world = world
+    Global.world = world
 
 
     const gameLoop = function () {
-        tick(Game)
+        tick(Global)
         requestAnimationFrame(gameLoop)
     }
