+++ examples/05-bloom/Global.js
