+++ examples/05-bloom/index.html
@@ -63,7 +63,7 @@
         </div>
 
 <script type="module">
-import Game             from './Game.js'
+import Global             from './Global.js'
 import * as Cobalt      from '../../bundle.js'
 import constants        from './constants.js'
 import dat              from 'https://cdn.skypack.dev/pin/dat.gui@v0.7.9-2wtQAdFH5SRwnJLDWGNz/mode=imports,min/optimized/dat.gui.js'
@@ -80,10 +80,10 @@ async function main () {
 
     const viewportWidth = constants.GAME_WIDTH
     const viewportHeight = constants.GAME_HEIGHT
-    Game.renderer = await Cobalt.init(canvas, viewportWidth, viewportHeight)
+    Global.renderer = await Cobalt.init(canvas, viewportWidth, viewportHeight)
 
     // instantiate all resource nodes
-    const hdrTex = await Cobalt.initNode(Game.renderer, {
+    const hdrTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -95,7 +95,7 @@ async function main () {
         }
     })
 
-    const spritesheet = await Cobalt.initNode(Game.renderer, {
+    const spritesheet = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:spritesheet',
         refs: { },
         options: {
@@ -105,7 +105,7 @@ async function main () {
         }
     })
 
-    const emissiveTex = await Cobalt.initNode(Game.renderer, {
+    const emissiveTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -117,7 +117,7 @@ async function main () {
         }
     })
 
-    const bloomTex = await Cobalt.initNode(Game.renderer, {
+    const bloomTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -130,7 +130,7 @@ async function main () {
     })
 
 
-    const spriteNode = await Cobalt.initNode(Game.renderer, {
+    const spriteNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:spriteHDR',
         refs: {
             // key is the var name defined in this node
@@ -145,7 +145,7 @@ async function main () {
     })
 
 
-    const bloomNode = await Cobalt.initNode(Game.renderer, {
+    const bloomNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:bloom',
         refs: {
             // key is the var name defined in this node
@@ -173,7 +173,7 @@ async function main () {
         }
     })
     
-    const tmpTex = await Cobalt.initNode(Game.renderer, {
+    const tmpTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -185,7 +185,7 @@ async function main () {
         }
     })
 
-    const compositeNode = await Cobalt.initNode(Game.renderer, {
+    const compositeNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:composite',
         refs: {
             hdr: hdrTex,
@@ -196,7 +196,7 @@ async function main () {
     })
 
 
-    const b = await Cobalt.initNode(Game.renderer, {
+    const b = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbBlit',
         refs: {
             in: tmpTex,
@@ -208,11 +208,11 @@ async function main () {
 
 
     // use resizeViewport to init values on load:
-    resizeViewport(Game.renderer, window.innerWidth, window.innerHeight)
+    resizeViewport(Global.renderer, window.innerWidth, window.innerHeight)
 
     // window resize is *expensive* - best to debounce:
     const debouncedResize = debounce(function () {
-       resizeViewport(Game.renderer, window.innerWidth, window.innerHeight) 
+       resizeViewport(Global.renderer, window.innerWidth, window.innerHeight) 
     }, 50)
 
     window.addEventListener('resize', debouncedResize, { passive: true })
@@ -238,7 +238,7 @@ async function main () {
     })
 
     const updateFinalParams = function (bloom_intensity=40.0, bloom_combine_constant=0.68, bloom_knee=0.2, bloom_threshold=0.1) {
-        const { device } = Game.renderer
+        const { device } = Global.renderer
         
         device.queue.writeBuffer(
             compositeNode.data.params_buf,
@@ -263,9 +263,9 @@ async function main () {
 
     
     const world = ECS.createWorld()
-    ECS.addSystem(world, rendererSystem(Game.renderer))
+    ECS.addSystem(world, rendererSystem(Global.renderer))
 
-    Cobalt.setViewportPosition(Game.renderer, [ 240, 135 ])
+    Cobalt.setViewportPosition(Global.renderer, [ 240, 135 ])
     
     ECS.addSystem(world, function cameraLoopSystem(world) {
         const onUpdate = function (dt) {
@@ -273,16 +273,16 @@ async function main () {
             const x = (Math.sin(elapsed / 2000) * 0.5 + 0.5) * 128
             const y = (Math.sin(elapsed / 5000) * 0.5 + 0.5) * 170
 
-            //const { tileScale } = Game.renderer.tile
+            //const { tileScale } = Global.renderer.tile
 
             //console.log(x, y)
-            Cobalt.setViewportPosition(Game.renderer, [ x, y ])
+            Cobalt.setViewportPosition(Global.renderer, [ x, y ])
         }
 
         return { onUpdate }
     })
 
-    Game.world = world
+    Global.world = world
 
     // test sprites
     const s2 = spriteEntity(world, {
@@ -303,7 +303,7 @@ async function main () {
 
 
     const gameLoop = function () {
-        tick(Game)
+        tick(Global)
         requestAnimationFrame(gameLoop)
     }
