+++ examples/06-displacement/Global.js
