+++ examples/08-light/Global.js
