+++ examples/08-light/index.html
@@ -60,7 +60,7 @@
         </div>
 
 <script type="module">
-import Game        from './Game.js'
+import Global      from './Global.js'
 import * as Cobalt from '../../bundle.js'
 import constants   from './constants.js'
 import dat         from 'https://cdn.skypack.dev/pin/dat.gui@v0.7.9-2wtQAdFH5SRwnJLDWGNz/mode=imports,min/optimized/dat.gui.js'
@@ -76,13 +76,13 @@ async function main () {
 
     const viewportWidth = constants.GAME_WIDTH
     const viewportHeight = constants.GAME_HEIGHT
-    Game.renderer = await Cobalt.init(canvas, viewportWidth, viewportHeight)
-    Game.renderer.viewport.zoom = 1;
+    Global.renderer = await Cobalt.init(canvas, viewportWidth, viewportHeight)
+    Global.renderer.viewport.zoom = 1;
 
     const viewportCenter = [0, 0];
 
     // instantiate all resource nodes
-    const tileAtlasNode = await Cobalt.initNode(Game.renderer, {
+    const tileAtlasNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:tileAtlas',
         refs: { },
         options: {
@@ -93,7 +93,7 @@ async function main () {
         }
     })
 
-    const hdrTex = await Cobalt.initNode(Game.renderer, {
+    const hdrTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -105,7 +105,7 @@ async function main () {
         }
     })
 
-    const lightCompositeHdrTex = await Cobalt.initNode(Game.renderer, {
+    const lightCompositeHdrTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -117,7 +117,7 @@ async function main () {
         }
     })
 
-    const emissiveTex = await Cobalt.initNode(Game.renderer, {
+    const emissiveTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -129,7 +129,7 @@ async function main () {
         }
     })
 
-    const bloomTex = await Cobalt.initNode(Game.renderer, {
+    const bloomTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -141,7 +141,7 @@ async function main () {
         }
     })
 
-    const bloomNode = await Cobalt.initNode(Game.renderer, {
+    const bloomNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:bloom',
         refs: {
             // key is the var name defined in this node
@@ -169,7 +169,7 @@ async function main () {
         }
     })
 
-    const tmpTex = await Cobalt.initNode(Game.renderer, {
+    const tmpTex = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbTexture',
         refs: { },
         options: {
@@ -182,8 +182,8 @@ async function main () {
     })
 
     // instantiate all runnable nodes
-    const tileNode = await Cobalt.initNode(Game.renderer, {
-        type: 'cobalt:tile',
+    const tileNode = await Cobalt.initNode(Global.renderer, {
+        type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
             // value is the var name in the cobalt resources dictionary
@@ -197,8 +197,8 @@ async function main () {
         }
     })
 
-    const tileNode2 = await Cobalt.initNode(Game.renderer, {
-        type: 'cobalt:tile',
+    const tileNode2 = await Cobalt.initNode(Global.renderer, {
+        type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
             // value is the var name in the cobalt resources dictionary
@@ -216,20 +216,20 @@ async function main () {
     const convertToViewportPosition = (clientPosition /* [number, number] */) /* [number, number] */ => {
         const canvasBox = canvas.getBoundingClientRect();
         return [
-             ((clientPosition[0] - canvasBox.left) / canvasBox.width) * constants.GAME_WIDTH  /Game.renderer.viewport.zoom,
-             ((clientPosition[1] - canvasBox.top) / canvasBox.height) * constants.GAME_HEIGHT /Game.renderer.viewport.zoom,
+             ((clientPosition[0] - canvasBox.left) / canvasBox.width) * constants.GAME_WIDTH  /Global.renderer.viewport.zoom,
+             ((clientPosition[1] - canvasBox.top) / canvasBox.height) * constants.GAME_HEIGHT /Global.renderer.viewport.zoom,
         ];
     }
 
     const convertToWorldPosition = (viewportPosition /* [number, number] */) /* [number, number] */ => {
         return [
-            viewportPosition[0] + viewportCenter[0] - 0.5 * constants.GAME_WIDTH / Game.renderer.viewport.zoom,
-            viewportPosition[1] + viewportCenter[1] - 0.5 * constants.GAME_HEIGHT/ Game.renderer.viewport.zoom,
+            viewportPosition[0] + viewportCenter[0] - 0.5 * constants.GAME_WIDTH / Global.renderer.viewport.zoom,
+            viewportPosition[1] + viewportCenter[1] - 0.5 * constants.GAME_HEIGHT/ Global.renderer.viewport.zoom,
         ];
     };
 
     // lights and shadows node
-    const light = await Cobalt.initNode(Game.renderer, {
+    const light = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:light',
         refs: {
             in: hdrTex,
@@ -298,7 +298,7 @@ async function main () {
             viewportCenter[0] -= movement[0];
             viewportCenter[1] -= movement[1];
 
-            Cobalt.setViewportPosition(Game.renderer, viewportCenter)
+            Cobalt.setViewportPosition(Global.renderer, viewportCenter)
         }
 
         if (temporaryObstacleSegment) {
@@ -345,7 +345,7 @@ async function main () {
     ];
     light.setLights(lights)
 
-    const compositeNode = await Cobalt.initNode(Game.renderer, {
+    const compositeNode = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:composite',
         refs: {
             hdr: lightCompositeHdrTex,
@@ -356,7 +356,7 @@ async function main () {
     })
 
 
-    const b = await Cobalt.initNode(Game.renderer, {
+    const b = await Cobalt.initNode(Global.renderer, {
         type: 'cobalt:fbBlit',
         refs: {
             in: tmpTex,
@@ -367,11 +367,11 @@ async function main () {
 
 
     // use resizeViewport to init values on load:
-    resizeViewport(Game.renderer, window.innerWidth, window.innerHeight)
+    resizeViewport(Global.renderer, window.innerWidth, window.innerHeight)
 
     // window resize is *expensive* - best to debounce:
     const debouncedResize = debounce(function () {
-       resizeViewport(Game.renderer, window.innerWidth, window.innerHeight) 
+       resizeViewport(Global.renderer, window.innerWidth, window.innerHeight) 
     }, 50)
 
     window.addEventListener('resize', debouncedResize, { passive: true })
@@ -396,13 +396,13 @@ async function main () {
             ];
             light.setLights(lights)
 
-            Cobalt.draw(Game.renderer)
+            Cobalt.draw(Global.renderer)
         }
 
         return { onUpdate }
     })
 
-    Cobalt.setViewportPosition(Game.renderer, viewportCenter)
+    Cobalt.setViewportPosition(Global.renderer, viewportCenter)
     
     ECS.addSystem(world, function cameraLoopSystem(world) {
         const onUpdate = function (dt) {
@@ -410,17 +410,17 @@ async function main () {
             // const x = (Math.sin(elapsed / 2000) * 0.5 + 0.5) * 128
             // const y = (Math.sin(elapsed / 5000) * 0.5 + 0.5) * 170
 
-            // Cobalt.setViewportPosition(Game.renderer, [ x, y ])
+            // Cobalt.setViewportPosition(Global.renderer, [ x, y ])
         }
 
         return { onUpdate }
     })
 
-    Game.world = world
+    Global.world = world
 
 
     const gameLoop = function () {
-        tick(Game)
+        tick(Global)
         requestAnimationFrame(gameLoop)
     }
